Changes for closing flyouts
